+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Environment variable `AZUREAUTH_CACHE_FILE` and option `--cache` to support a custom cache location on Windows.
+- Added Integrated Windows Authentication functionality as one of the auth flows.
+
 ## [0.3.1] - 2022-06-07
 ### Fixed
 - Fixed a bug where the tenant and resource ids were swapped in the telemetry events.